+++ shared/project-links.ts
@@ -1,0 +1,3 @@
+export const FACTORY_CAREERS_REPOSITORY_URL = 'https://github.com/theFactoryHQ/factory-careers'
+export const FACTORY_CAREERS_NEW_ISSUE_URL = `${FACTORY_CAREERS_REPOSITORY_URL}/issues/new`
+export const FACTORY_CAREERS_RELEASES_URL = `${FACTORY_CAREERS_REPOSITORY_URL}/releases`